refactor(Filter): move allTypes array into its own file in util and import into Filter
